+++ src/Tests/Dfe.Complete.Domain.Tests/Enums/NoteTaskIdentifierTests.cs
@@ -16,6 +16,8 @@ namespace Dfe.Complete.Domain.Tests.Enums
         [InlineData(NoteTaskIdentifier.DeedOfVariation, "deed_of_variation", "Deed of variation")]
         [InlineData(NoteTaskIdentifier.RedactAndSendDocuments, "redact_and_send_documents", "Redact and send documents")]
         [InlineData(NoteTaskIdentifier.ConfirmProposedCapacityOfTheAcademy, "proposed_capacity_of_the_academy", "Confirm the proposed capacity of the academy")]
+        [InlineData(NoteTaskIdentifier.DeclarationOfExpenditureCertificate, "receive_declaration_of_expenditure_certificate", "Receive declaration of expenditure certificate")]
+        [InlineData(NoteTaskIdentifier.ConfirmAllConditionsMet, "confirm_all_conditions_has_been_met", "Confirm all conditions have been met")]
         public void NoteTaskIdentifier_ShouldHaveCorrectDescription(NoteTaskIdentifier identifier, string expectedDescription, string expectedDisplayExpectation)
         {
             // Act